# identity_service/docs/swagger.yaml
definitions:
  github_com_windevkay_flho_identity_service_internal_data.Identity:
    properties:
      activated:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_windevkay_flho_identity_service_internal_services.ActivateIdentityInput:
    properties:
      token:
        type: string
    type: object
  github_com_windevkay_flho_identity_service_internal_services.RegisterIdentityInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  helpers.Envelope:
    additionalProperties: true
    type: object
info:
  contact: {}
paths:
  /v1/auth/token:
    post:
      consumes:
      - application/json
      description: Create a JWT token for authentication
      parameters:
      - description: Authentication credentials
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                authentication_token:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  type: string
              type: object
      summary: Create authentication token
      tags:
      - auth
  /v1/users:
    post:
      consumes:
      - application/json
      description: Register a new identity in the system
      parameters:
      - description: Identity registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_windevkay_flho_identity_service_internal_services.RegisterIdentityInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                identity:
                  $ref: '#/definitions/github_com_windevkay_flho_identity_service_internal_data.Identity'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  type: string
              type: object
      summary: Register new identity
      tags:
      - identity
  /v1/users/activate:
    put:
      consumes:
      - application/json
      description: Activate a registered identity
      parameters:
      - description: Identity activation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_windevkay_flho_identity_service_internal_services.ActivateIdentityInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                identity:
                  $ref: '#/definitions/github_com_windevkay_flho_identity_service_internal_data.Identity'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.Envelope'
            - properties:
                error:
                  type: string
              type: object
      summary: Activate identity
      tags:
      - identity
swagger: "2.0"